Added some sets of checks to the connection controller for unauthorized access. Need to write an exception handler so these are better handled

+++ src/DataCat/Controllers/DataConnectionController.cs
@@ -25,8 +25,10 @@
         {
             try
             {
-                // TODO: Getting any connection like it doesn't matter. 
                 var result = await service.Find(id);
+                if (result.User != this.User.GetUserId())
+                    return Unauthorized();
+
                 return Ok(result);
             }
             catch (EntityNotFoundException ex)
@@ -52,6 +54,10 @@
         {
             try
             {
+                var connection = await service.Find(id);
+                if (connection.User != this.User.GetUserId())
+                    return Unauthorized();
+
                 // TODO: Deleting any connection like it doesn't matter.
                 var result = await service.Delete(id);
                 return Ok(result);
@@ -69,7 +75,10 @@
             if (model == null || !ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            // TODO: Updating any connection like it doesn't matter.
+            var connection = await service.Find(id);
+            if (connection.User != this.User.GetUserId())
+                return Unauthorized();
+
             var result = await service.Update(model.ToEntity(id, this.User.GetUserId()));
             return Ok(result);
         }
